model works great, may be try to increase performance further
