+++ functions.py
@@ -1,27 +1,301 @@
 # Imports
-import os
-import sys
-import gc
-import random
+import os, sys, gc, random, time, warnings, pickle
 from datetime import datetime
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import seaborn as sns
-import xgboost as xgb
-from xgboost import plot_importance
-from sklearn.metrics import r2_score
+from contextlib import contextmanager
 from sklearn.decomposition import KernelPCA, PCA
 from sklearn.manifold import TSNE
-from sklearn.utils import shuffle
-from sklearn.model_selection import train_test_split, RandomizedSearchCV
+from sklearn.linear_model import Lasso
+from sklearn.model_selection import GridSearchCV
+from sklearn.model_selection import train_test_split
+from sklearn.feature_selection import SelectFromModel
+from sklearn.metrics import roc_auc_score, roc_curve
+from sklearn.model_selection import KFold, StratifiedKFold
 from sklearn.impute import SimpleImputer
 from sklearn.preprocessing import LabelEncoder, OrdinalEncoder, StandardScaler, PolynomialFeatures
 
+
 # -------------------------------------------------------------------------------------------------
+@contextmanager
+def timer(title):
+    t0 = time.time()
+    yield
+    print("{} - done in {:.0f}s".format(title, time.time() - t0))
+
+# One-hot encoding for categorical columns with get_dummies
+
+
+def one_hot_encoder(df, nan_as_category=True):
+    original_columns = list(df.columns)
+    categorical_columns = [
+        col for col in df.columns if df[col].dtype == 'object']
+    df = pd.get_dummies(df, columns=categorical_columns,
+                        dummy_na=nan_as_category)
+    new_columns = [c for c in df.columns if c not in original_columns]
+    return df, new_columns
+
+# Preprocess application_train.csv and application_test.csv
+
+
+def application_train_test(num_rows=None, nan_as_category=False):
+    # Read data and merge
+    df = pd.read_csv('data/application_train.csv', nrows=num_rows)
+
+    print("Train samples: {}".format(len(df)))
+
+    # Optional: Remove 4 applications with XNA CODE_GENDER (train set)
+    df = df[df['CODE_GENDER'] != 'XNA']
+
+    # Categorical features with Binary encode (0 or 1; two categories)
+    for bin_feature in ['CODE_GENDER', 'FLAG_OWN_CAR', 'FLAG_OWN_REALTY']:
+        df[bin_feature], uniques = pd.factorize(df[bin_feature])
+    # Categorical features with One-Hot encode
+    df, cat_cols = one_hot_encoder(df, nan_as_category)
+
+    # NaN values for DAYS_EMPLOYED: 365.243 -> nan
+    df['DAYS_EMPLOYED'].replace(365243, np.nan, inplace=True)
+    # Some simple new features (percentages)
+    df['DAYS_EMPLOYED_PERC'] = df['DAYS_EMPLOYED'] / df['DAYS_BIRTH']
+    df['INCOME_CREDIT_PERC'] = df['AMT_INCOME_TOTAL'] / df['AMT_CREDIT']
+    df['INCOME_PER_PERSON'] = df['AMT_INCOME_TOTAL'] / df['CNT_FAM_MEMBERS']
+    df['ANNUITY_INCOME_PERC'] = df['AMT_ANNUITY'] / df['AMT_INCOME_TOTAL']
+    df['PAYMENT_RATE'] = df['AMT_ANNUITY'] / df['AMT_CREDIT']
+
+    return df
+
+# Preprocess bureau.csv and bureau_balance.csv
+
+
+def bureau_and_balance(num_rows=None, nan_as_category=True):
+    bureau = pd.read_csv('data/bureau.csv', nrows=num_rows)
+    bb = pd.read_csv('data/bureau_balance.csv', nrows=num_rows)
+    bb, bb_cat = one_hot_encoder(bb, nan_as_category)
+    bureau, bureau_cat = one_hot_encoder(bureau, nan_as_category)
+
+    # Bureau balance: Perform aggregations and merge with bureau.csv
+    bb_aggregations = {'MONTHS_BALANCE': ['min', 'max', 'size']}
+    for col in bb_cat:
+        bb_aggregations[col] = ['mean']
+    bb_agg = bb.groupby('SK_ID_BUREAU').agg(bb_aggregations)
+    bb_agg.columns = pd.Index([e[0] + "_" + e[1].upper()
+                               for e in bb_agg.columns.tolist()])
+    bureau = bureau.join(bb_agg, how='left', on='SK_ID_BUREAU')
+    bureau.drop(['SK_ID_BUREAU'], axis=1, inplace=True)
+    del bb, bb_agg
+    gc.collect()
+
+    # Bureau and bureau_balance numeric features
+    num_aggregations = {
+        'DAYS_CREDIT': ['min', 'max', 'mean', 'var'],
+        'DAYS_CREDIT_ENDDATE': ['min', 'max', 'mean'],
+        'DAYS_CREDIT_UPDATE': ['mean'],
+        'CREDIT_DAY_OVERDUE': ['max', 'mean'],
+        'AMT_CREDIT_MAX_OVERDUE': ['mean'],
+        'AMT_CREDIT_SUM': ['max', 'mean', 'sum'],
+        'AMT_CREDIT_SUM_DEBT': ['max', 'mean', 'sum'],
+        'AMT_CREDIT_SUM_OVERDUE': ['mean'],
+        'AMT_CREDIT_SUM_LIMIT': ['mean', 'sum'],
+        'AMT_ANNUITY': ['max', 'mean'],
+        'CNT_CREDIT_PROLONG': ['sum'],
+        'MONTHS_BALANCE_MIN': ['min'],
+        'MONTHS_BALANCE_MAX': ['max'],
+        'MONTHS_BALANCE_SIZE': ['mean', 'sum']
+    }
+    # Bureau and bureau_balance categorical features
+    cat_aggregations = {}
+    for cat in bureau_cat:
+        cat_aggregations[cat] = ['mean']
+    for cat in bb_cat:
+        cat_aggregations[cat + "_MEAN"] = ['mean']
+
+    bureau_agg = bureau.groupby('SK_ID_CURR').agg(
+        {**num_aggregations, **cat_aggregations})
+    bureau_agg.columns = pd.Index(
+        ['BURO_' + e[0] + "_" + e[1].upper() for e in bureau_agg.columns.tolist()])
+    # Bureau: Active credits - using only numerical aggregations
+    active = bureau[bureau['CREDIT_ACTIVE_Active'] == 1]
+    active_agg = active.groupby('SK_ID_CURR').agg(num_aggregations)
+    active_agg.columns = pd.Index(
+        ['ACTIVE_' + e[0] + "_" + e[1].upper() for e in active_agg.columns.tolist()])
+    bureau_agg = bureau_agg.join(active_agg, how='left', on='SK_ID_CURR')
+    del active, active_agg
+    gc.collect()
+    # Bureau: Closed credits - using only numerical aggregations
+    closed = bureau[bureau['CREDIT_ACTIVE_Closed'] == 1]
+    closed_agg = closed.groupby('SK_ID_CURR').agg(num_aggregations)
+    closed_agg.columns = pd.Index(
+        ['CLOSED_' + e[0] + "_" + e[1].upper() for e in closed_agg.columns.tolist()])
+    bureau_agg = bureau_agg.join(closed_agg, how='left', on='SK_ID_CURR')
+    del closed, closed_agg, bureau
+    gc.collect()
+    return bureau_agg
+
+# Preprocess previous_applications.csv
+
+
+def previous_applications(num_rows=None, nan_as_category=True):
+    prev = pd.read_csv('data/previous_application.csv', nrows=num_rows)
+    prev, cat_cols = one_hot_encoder(prev, nan_as_category=True)
+    # Days 365.243 values -> nan
+    prev['DAYS_FIRST_DRAWING'].replace(365243, np.nan, inplace=True)
+    prev['DAYS_FIRST_DUE'].replace(365243, np.nan, inplace=True)
+    prev['DAYS_LAST_DUE_1ST_VERSION'].replace(365243, np.nan, inplace=True)
+    prev['DAYS_LAST_DUE'].replace(365243, np.nan, inplace=True)
+    prev['DAYS_TERMINATION'].replace(365243, np.nan, inplace=True)
+    # Add feature: value ask / value received percentage
+    prev['APP_CREDIT_PERC'] = prev['AMT_APPLICATION'] / prev['AMT_CREDIT']
+    # Previous applications numeric features
+    num_aggregations = {
+        'AMT_ANNUITY': ['min', 'max', 'mean'],
+        'AMT_APPLICATION': ['min', 'max', 'mean'],
+        'AMT_CREDIT': ['min', 'max', 'mean'],
+        'APP_CREDIT_PERC': ['min', 'max', 'mean', 'var'],
+        'AMT_DOWN_PAYMENT': ['min', 'max', 'mean'],
+        'AMT_GOODS_PRICE': ['min', 'max', 'mean'],
+        'HOUR_APPR_PROCESS_START': ['min', 'max', 'mean'],
+        'RATE_DOWN_PAYMENT': ['min', 'max', 'mean'],
+        'DAYS_DECISION': ['min', 'max', 'mean'],
+        'CNT_PAYMENT': ['mean', 'sum'],
+    }
+    # Previous applications categorical features
+    cat_aggregations = {}
+    for cat in cat_cols:
+        cat_aggregations[cat] = ['mean']
+
+    prev_agg = prev.groupby('SK_ID_CURR').agg(
+        {**num_aggregations, **cat_aggregations})
+    prev_agg.columns = pd.Index(
+        ['PREV_' + e[0] + "_" + e[1].upper() for e in prev_agg.columns.tolist()])
+    # Previous Applications: Approved Applications - only numerical features
+    approved = prev[prev['NAME_CONTRACT_STATUS_Approved'] == 1]
+    approved_agg = approved.groupby('SK_ID_CURR').agg(num_aggregations)
+    approved_agg.columns = pd.Index(
+        ['APPROVED_' + e[0] + "_" + e[1].upper() for e in approved_agg.columns.tolist()])
+    prev_agg = prev_agg.join(approved_agg, how='left', on='SK_ID_CURR')
+    # Previous Applications: Refused Applications - only numerical features
+    refused = prev[prev['NAME_CONTRACT_STATUS_Refused'] == 1]
+    refused_agg = refused.groupby('SK_ID_CURR').agg(num_aggregations)
+    refused_agg.columns = pd.Index(
+        ['REFUSED_' + e[0] + "_" + e[1].upper() for e in refused_agg.columns.tolist()])
+    prev_agg = prev_agg.join(refused_agg, how='left', on='SK_ID_CURR')
+    del refused, refused_agg, approved, approved_agg, prev
+    gc.collect()
+    return prev_agg
+
+# Preprocess POS_CASH_balance.csv
+
+
+def pos_cash(num_rows=None, nan_as_category=True):
+    pos = pd.read_csv('data/POS_CASH_balance.csv', nrows=num_rows)
+    pos, cat_cols = one_hot_encoder(pos, nan_as_category=True)
+    # Features
+    aggregations = {
+        'MONTHS_BALANCE': ['max', 'mean', 'size'],
+        'SK_DPD': ['max', 'mean'],
+        'SK_DPD_DEF': ['max', 'mean']
+    }
+    for cat in cat_cols:
+        aggregations[cat] = ['mean']
+
+    pos_agg = pos.groupby('SK_ID_CURR').agg(aggregations)
+    pos_agg.columns = pd.Index(
+        ['POS_' + e[0] + "_" + e[1].upper() for e in pos_agg.columns.tolist()])
+    # Count pos cash accounts
+    pos_agg['POS_COUNT'] = pos.groupby('SK_ID_CURR').size()
+    del pos
+    gc.collect()
+    return pos_agg
+
+# Preprocess installments_payments.csv
+
+
+def installments_payments(num_rows=None, nan_as_category=True):
+    ins = pd.read_csv('data/installments_payments.csv', nrows=num_rows)
+    ins, cat_cols = one_hot_encoder(ins, nan_as_category=True)
+    # Percentage and difference paid in each installment (amount paid and installment value)
+    ins['PAYMENT_PERC'] = ins['AMT_PAYMENT'] / ins['AMT_INSTALMENT']
+    ins['PAYMENT_DIFF'] = ins['AMT_INSTALMENT'] - ins['AMT_PAYMENT']
+    # Days past due and days before due (no negative values)
+    ins['DPD'] = ins['DAYS_ENTRY_PAYMENT'] - ins['DAYS_INSTALMENT']
+    ins['DBD'] = ins['DAYS_INSTALMENT'] - ins['DAYS_ENTRY_PAYMENT']
+    ins['DPD'] = ins['DPD'].apply(lambda x: x if x > 0 else 0)
+    ins['DBD'] = ins['DBD'].apply(lambda x: x if x > 0 else 0)
+    # Features: Perform aggregations
+    aggregations = {
+        'NUM_INSTALMENT_VERSION': ['nunique'],
+        'DPD': ['max', 'mean', 'sum'],
+        'DBD': ['max', 'mean', 'sum'],
+        'PAYMENT_PERC': ['max', 'mean', 'sum', 'var'],
+        'PAYMENT_DIFF': ['max', 'mean', 'sum', 'var'],
+        'AMT_INSTALMENT': ['max', 'mean', 'sum'],
+        'AMT_PAYMENT': ['min', 'max', 'mean', 'sum'],
+        'DAYS_ENTRY_PAYMENT': ['max', 'mean', 'sum']
+    }
+    for cat in cat_cols:
+        aggregations[cat] = ['mean']
+    ins_agg = ins.groupby('SK_ID_CURR').agg(aggregations)
+    ins_agg.columns = pd.Index(
+        ['INSTAL_' + e[0] + "_" + e[1].upper() for e in ins_agg.columns.tolist()])
+    # Count installments accounts
+    ins_agg['INSTAL_COUNT'] = ins.groupby('SK_ID_CURR').size()
+    del ins
+    gc.collect()
+    return ins_agg
+
+# Preprocess credit_card_balance.csv
+
+
+def credit_card_balance(num_rows=None, nan_as_category=True):
+    cc = pd.read_csv('data/credit_card_balance.csv', nrows=num_rows)
+    cc, cat_cols = one_hot_encoder(cc, nan_as_category=True)
+    # General aggregations
+    cc.drop(['SK_ID_PREV'], axis=1, inplace=True)
+    cc_agg = cc.groupby('SK_ID_CURR').agg(['min', 'max', 'mean', 'sum', 'var'])
+    cc_agg.columns = pd.Index(['CC_' + e[0] + "_" + e[1].upper()
+                               for e in cc_agg.columns.tolist()])
+    # Count credit card lines
+    cc_agg['CC_COUNT'] = cc.groupby('SK_ID_CURR').size()
+    del cc
+    gc.collect()
+    return cc_agg
+
+
+def display_importances(feature_importance_df_):
+    cols = feature_importance_df_[["feature", "importance"]].groupby(
+        "feature").mean().sort_values(by="importance", ascending=False)[:40].index
+    best_features = feature_importance_df_.loc[feature_importance_df_.feature.isin(
+        cols)]
+    plt.figure(figsize=(8, 10))
+    sns.barplot(x="importance", y="feature", data=best_features.sort_values(
+        by="importance", ascending=False))
+    plt.title('LightGBM Features (avg over folds)')
+    plt.tight_layout()
+    plt.savefig('lgbm_importances01.png')
+
+
+def outliers_transformer(df):
+    """ Take dataframe and transform all outliers of each column to the median value of that column """
+
+    # list of columns names
+    columns_names = list(df.columns)
+
+    # transform each column outliers to the median value of that column
+    for col in columns_names:
+        if col != 'TARGET':
+            median = df[col].median()
+            std = df[col].std()
+
+            df.loc[((df[col] - median).abs() > std) == True, col] = np.nan
+
+            df[col].fillna(median, inplace=True)
+
+    return df
+
+
 # Function to calculate missing values by column# Funct
-
-
 def missing_values_table(df):
     # Total missing values
     mis_val = df.isnull().sum()
@@ -45,7 +319,10 @@ def missing_values_table(df):
     print("Your selected dataframe has " + str(df.shape[1]) + " columns.\n"
           "There are " + str(mis_val_table_ren_columns.shape[0]) +
           " columns that have missing values.")
-
+    # free memory
+    gc.enable()
+    del mis_val, mis_val_percent, mis_val_table
+    gc.collect()
     # Return the dataframe with missing information
     return mis_val_table_ren_columns
 # -------------------------------------------------------------------------------------------------
@@ -346,7 +623,7 @@ def convert_types(df, print_info=False):
     return df
 
 
-def remove_missing_columns(train, threshold=90):
+def remove_missing_columns(train, threshold=80):
     # Calculate missing stats for train and test (remember to calculate a percent!)
     train_miss = pd.DataFrame(train.isnull().sum())
     train_miss['percent'] = 100 * train_miss[0] / len(train)
@@ -435,20 +712,20 @@ def aggregate_client(df, group_vars, df_names):
     gc.collect()
 
     return df_by_client
-# print out importances features (bar plot horiz)
 
 
-def print_importance_features(model, index):
-    # Create a pd.Series of features importances
-    importances = pd.Series(data=model.feature_importances_, index=index)
-
-    # Sort importances
-    importances_sorted = importances.sort_values()
-    importances_sorted_toplot = importances_sorted[importances_sorted >= 0.5 *  # trick to print just some most important features
-                                                   importances.mean()]
-
-    # Draw a horizontal barplot of importances_sorted
-    importances_sorted_toplot.plot(kind='barh', color='lightgreen')
-    plt.title('Features Importances')
-    plt.show()
-    return importances
+def print_importance_features(model, index, num_features=5):
+    """
+    Parameters
+    --------
+    model: model used for prediction
+    index: data.columns (name of the columns)
+    num_features: top number of features to show
+    
+    Return
+    --------
+    return dataframe with sorted features importances higher than the mean 
+     """
+    top_features = pd.DataFrame({'features': index, 'importance': model.feature_importances_}).sort_values('importance', ascending=False).head(num_features)
+    sns.barplot(x=top_features['importance'], y=top_features['features'])
+    return top_features
