--- p7_01_functions.py
@@ -1,458 +1,0 @@
-# Imports
-import os
-import sys
-import gc
-import random
-from datetime import datetime
-import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
-import seaborn as sns
-from sklearn.decomposition import KernelPCA, PCA
-from sklearn.manifold import TSNE
-from sklearn.linear_model import Lasso
-from sklearn.model_selection import train_test_split
-from sklearn.feature_selection import SelectFromModel
-from sklearn.impute import SimpleImputer
-from sklearn.preprocessing import LabelEncoder, OrdinalEncoder, StandardScaler, PolynomialFeatures
-
-
-# -------------------------------------------------------------------------------------------------
-# Function to calculate missing values by column# Funct
-def missing_values_table(df):
-    # Total missing values
-    mis_val = df.isnull().sum()
-
-    # Percentage of missing values
-    mis_val_percent = 100 * df.isnull().sum() / len(df)
-
-    # Make a table with the results
-    mis_val_table = pd.concat([mis_val, mis_val_percent], axis=1)
-
-    # Rename the columns
-    mis_val_table_ren_columns = mis_val_table.rename(
-        columns={0: 'Missing Values', 1: '% of Total Missing Values'})
-
-    # Sort the table by percentage of missing descending
-    mis_val_table_ren_columns = mis_val_table_ren_columns[
-        mis_val_table_ren_columns.iloc[:, 1] != 0].sort_values(
-        '% of Total Missing Values', ascending=False).round(1)
-
-    # Print some summary information
-    print("Your selected dataframe has " + str(df.shape[1]) + " columns.\n"
-          "There are " + str(mis_val_table_ren_columns.shape[0]) +
-          " columns that have missing values.")
-    # free memory 
-    gc.enable()
-    del mis_val, mis_val_percent, mis_val_table
-    gc.collect()
-    # Return the dataframe with missing information
-    return mis_val_table_ren_columns
-# -------------------------------------------------------------------------------------------------
-# Plots the disribution of a variable colored by value of the target
-
-
-def kde_target(var_name, df):
-
-    # Calculate the correlation coefficient between the new variable and the target
-    corr = df['TARGET'].corr(df[var_name])
-
-    # Calculate medians for repaid vs not repaid
-    avg_repaid = df.loc[df['TARGET'] == 0, var_name].median()
-    avg_not_repaid = df.loc[df['TARGET'] == 1, var_name].median()
-
-    plt.figure(figsize=(12, 6))
-
-    # Plot the distribution for target == 0 and target == 1
-    sns.kdeplot(df.loc[df['TARGET'] == 0, var_name], label='TARGET == 0')
-    sns.kdeplot(df.loc[df['TARGET'] == 1, var_name], label='TARGET == 1')
-
-    # label the plot
-    plt.xlabel(var_name)
-    plt.ylabel('Density')
-    plt.title('%s Distribution' % var_name)
-    plt.legend()
-
-    # print out the correlation
-    print('The correlation between %s and the TARGET is %0.4f' % (var_name, corr))
-    # Print out average values
-    print('Median value for loan that was not repaid = %0.4f' % avg_not_repaid)
-    print('Median value for loan that was repaid =     %0.4f' % avg_repaid)
-
-# -------------------------------------------------------------------------------------------------
-
-
-def agg_numeric(df, group_var, df_name):
-    """Aggregates the numeric values in a dataframe. This can
-    be used to create features for each instance of the grouping variable.
-
-    Parameters
-    --------
-        df (dataframe): 
-            the dataframe to calculate the statistics on
-        group_var (string): 
-            the variable by which to group df
-        df_name (string): 
-            the variable used to rename the columns
-
-    Return
-    --------
-        agg (dataframe): 
-            a dataframe with the statistics aggregated for 
-            all numeric columns. Each instance of the grouping variable will have 
-            the statistics (mean, min, max, sum; currently supported) calculated. 
-            The columns are also renamed to keep track of features created.
-
-    """
-    # Remove id variables other than grouping variable
-    for col in df:
-        if col != group_var and 'SK_ID' in col:
-            df = df.drop(columns=col)
-
-    group_ids = df[group_var]
-    numeric_df = df.select_dtypes('number')
-    numeric_df[group_var] = group_ids
-
-    # Group by the specified variable and calculate the statistics
-    agg = numeric_df.groupby(group_var).agg(
-        ['count', 'mean', 'max', 'min', 'sum']).reset_index()
-
-    # Need to create new column names
-    columns = [group_var]
-
-    # Iterate through the variables names
-    for var in agg.columns.levels[0]:
-        # Skip the grouping variable
-        if var != group_var:
-            # Iterate through the stat names
-            for stat in agg.columns.levels[1][:-1]:
-                # Make a new column name for the variable and stat
-                columns.append('%s_%s_%s' % (df_name, var, stat))
-
-    agg.columns = columns
-    agg = agg.reset_index()
-    return agg
-# -------------------------------------------------------------------------------------------------
-
-# -------------------------------------------------------------------------------------------------
-# Function to calculate correlations with the target for a dataframe
-
-
-def target_corrs(df):
-
-    # List of correlations
-    corrs = []
-
-    # Iterate through the columns
-    for col in df.columns:
-        print(col)
-        # Skip the target column
-        if col != 'TARGET':
-            # Calculate correlation with the target
-            corr = df['TARGET'].corr(df[col])
-
-            # Append the list as a tuple
-            corrs.append((col, corr))
-
-    # Sort by absolute magnitude of correlations
-    corrs = sorted(corrs, key=lambda x: abs(x[1]), reverse=True)
-
-    return corrs
-
-# -------------------------------------------------------------------------------------------------
-
-
-def count_categorical(df, group_var, df_name):
-    """Computes counts and normalized counts for each observation
-    of `group_var` of each unique category in every categorical variable
-
-    Parameters
-    --------
-    df : dataframe 
-        The dataframe to calculate the value counts for.
-
-    group_var : string
-        The variable by which to group the dataframe. For each unique
-        value of this variable, the final dataframe will have one row
-
-    df_name : string
-        Variable added to the front of column names to keep track of columns
-
-
-    Return
-    --------
-    categorical : dataframe
-        A dataframe with counts and normalized counts of each unique category in every categorical variable
-        with one row for every unique value of the `group_var`.
-
-    """
-
-    # Select the categorical columns
-    categorical = pd.get_dummies(df.select_dtypes('object'))
-
-    # Make sure to put the identifying id on the column
-    categorical[group_var] = df[group_var]
-
-    # Groupby the group var and calculate the sum and mean
-    categorical = categorical.groupby(group_var).agg(['sum', 'mean'])
-
-    column_names = []
-
-    # Iterate through the columns in level 0
-    for var in categorical.columns.levels[0]:
-        # Iterate through the stats in level 1
-        for stat in ['count', 'count_norm']:
-            # Make a new column name
-            column_names.append('%s_%s_%s' % (df_name, var, stat))
-
-    categorical.columns = column_names
-    categorical = categorical.reset_index()
-    return categorical
-
-
-def plot_feature_importances(df):
-    """
-    Plot importances returned by a model. This can work with any measure of
-    feature importance provided that higher importance is better. 
-
-    Args:
-        df (dataframe): feature importances. Must have the features in a column
-        called `features` and the importances in a column called `importance
-
-    Returns:
-        shows a plot of the 15 most importance features
-
-        df (dataframe): feature importances sorted by importance (highest to lowest) 
-        with a column for normalized importance
-        """
-
-    # Sort features according to importance
-    df = df.sort_values('importance', ascending=False).reset_index()
-
-    # Normalize the feature importances to add up to one
-    df['importance_normalized'] = df['importance'] / df['importance'].sum()
-
-    # Make a horizontal bar chart of feature importances
-    plt.figure(figsize=(10, 6))
-    ax = plt.subplot()
-
-    # Need to reverse the index to plot most important on top
-    ax.barh(list(reversed(list(df.index[:15]))),
-            df['importance_normalized'].head(15),
-            align='center', edgecolor='k')
-
-    # Set the yticks and labels
-    ax.set_yticks(list(reversed(list(df.index[:15]))))
-    ax.set_yticklabels(df['feature'].head(15))
-
-    # Plot labeling
-    plt.xlabel('Normalized Importance')
-    plt.title('Feature Importances')
-    plt.show()
-    return df
-
-
-def agg_categorical(df, group_var, df_name):
-    """
-    Aggregates the categorical features in a child dataframe
-    for each observation of the parent variable.
-
-    Parameters
-    --------
-    df : dataframe 
-        The dataframe to calculate the value counts for.
-
-    group_var : string
-        The variable by which to group and aggregate the dataframe. For each unique
-        value of this variable, the final dataframe will have one row
-
-    df_name : string
-        Variable added to the front of column names to keep track of columns
-
-
-    Return
-    --------
-    categorical : dataframe
-        A dataframe with aggregated statistics for each observation of the group_var
-        The columns are also renamed and columns with duplicate values are removed.
-
-    """
-
-    # Select the categorical columns
-    categorical = pd.get_dummies(df.select_dtypes('category'))
-
-    # Make sure to put the identifying id on the column
-    categorical[group_var] = df[group_var]
-
-    # Groupby the group var and calculate the sum and mean
-    categorical = categorical.groupby(group_var).agg(['sum', 'count', 'mean'])
-
-    column_names = []
-
-    # Iterate through the columns in level 0
-    for var in categorical.columns.levels[0]:
-        # Iterate through the stats in level 1
-        for stat in ['sum', 'count', 'mean']:
-            # Make a new column name
-            column_names.append('%s_%s_%s' % (df_name, var, stat))
-
-    categorical.columns = column_names
-
-    # Remove duplicate columns by values
-    _, idx = np.unique(categorical, axis=1, return_index=True)
-    categorical = categorical.iloc[:, idx]
-
-    return categorical
-
-
-def return_size(df):
-    """Return size of dataframe in gigabytes"""
-    return round(sys.getsizeof(df) / 1e9, 2)
-
-
-def convert_types(df, print_info=False):
-
-    original_memory = df.memory_usage().sum()
-
-    # Iterate through each column
-    for c in df:
-
-        # Convert ids and booleans to integers
-        if ('SK_ID' in c):
-            df[c] = df[c].fillna(0).astype(np.int32)
-
-        # Convert objects to category
-        elif (df[c].dtype == 'object') and (df[c].nunique() < df.shape[0]):
-            df[c] = df[c].astype('category')
-
-        # Booleans mapped to integers
-        elif list(df[c].unique()) == [1, 0]:
-            df[c] = df[c].astype(bool)
-
-        # Float64 to float32
-        elif df[c].dtype == float:
-            df[c] = df[c].astype(np.float32)
-
-        # Int64 to int32
-        elif df[c].dtype == int:
-            df[c] = df[c].astype(np.int32)
-
-    new_memory = df.memory_usage().sum()
-
-    if print_info:
-        print(f'Original Memory Usage: {round(original_memory / 1e9, 2)} gb.')
-        print(f'New Memory Usage: {round(new_memory / 1e9, 2)} gb.')
-
-    return df
-
-
-def remove_missing_columns(train, threshold=90):
-    # Calculate missing stats for train and test (remember to calculate a percent!)
-    train_miss = pd.DataFrame(train.isnull().sum())
-    train_miss['percent'] = 100 * train_miss[0] / len(train)
-
-    # list of missing columns for train and test
-    missing_train_columns = list(
-        train_miss.index[train_miss['percent'] > threshold])
-
-    # Combine the two lists together
-    missing_columns = list(set(missing_train_columns))
-
-    # Print information
-    print('There are %d columns with greater than %d%% missing values.' %
-          (len(missing_columns), threshold))
-
-    # Drop the missing columns and return
-    train = train.drop(columns=missing_columns)
-
-    return train
-
-
-def aggregate_client(df, group_vars, df_names):
-    """Aggregate a dataframe with data at the loan level 
-    at the client level
-
-    Args:
-        df (dataframe): data at the loan level
-        group_vars (list of two strings): grouping variables for the loan 
-        and then the client (example ['SK_ID_PREV', 'SK_ID_CURR'])
-        names (list of two strings): names to call the resulting columns
-        (example ['cash', 'client'])
-
-    Returns:
-        df_client (dataframe): aggregated numeric stats at the client level. 
-        Each client will have a single row with all the numeric data aggregated
-    """
-
-    # Aggregate the numeric columns
-    df_agg = agg_numeric(df, group_var=group_vars[0], df_name=df_names[0])
-
-    # If there are categorical variables
-    if any(df.dtypes == 'category'):
-
-        # Count the categorical columns
-        df_counts = agg_categorical(
-            df, group_var=group_vars[0], df_name=df_names[0])
-
-        # Merge the numeric and categorical
-        df_by_loan = df_counts.merge(df_agg, on=group_vars[0], how='outer')
-
-        gc.enable()
-        del df_agg, df_counts
-        gc.collect()
-
-        # Merge to get the client id in dataframe
-        df_by_loan = df_by_loan.merge(
-            df[[group_vars[0], group_vars[1]]], on=group_vars[0], how='left')
-
-        # Remove the loan id
-        df_by_loan = df_by_loan.drop(columns=[group_vars[0]])
-
-        # Aggregate numeric stats by column
-        df_by_client = agg_numeric(
-            df_by_loan, group_var=group_vars[1], df_name=df_names[1])
-
-    # No categorical variables
-    else:
-        # Merge to get the client id in dataframe
-        df_by_loan = df_agg.merge(
-            df[[group_vars[0], group_vars[1]]], on=group_vars[0], how='left')
-
-        gc.enable()
-        del df_agg
-        gc.collect()
-
-        # Remove the loan id
-        df_by_loan = df_by_loan.drop(columns=[group_vars[0]])
-
-        # Aggregate numeric stats by column
-        df_by_client = agg_numeric(
-            df_by_loan, group_var=group_vars[1], df_name=df_names[1])
-
-    # Memory management
-    gc.enable()
-    del df, df_by_loan
-    gc.collect()
-
-    return df_by_client
-
-
-def print_importance_features(model, index):
-    """
-    Return
-    --------
-    return dataframe with sorted features importances higher than the mean 
-     """
-    # Create a pd.Series of features importances
-    importances = pd.Series(data=model.feature_importances_, index=index)
-
-    # Sort importances
-    importances_sorted = importances.sort_values()
-    importances_sorted_toplot = importances_sorted[importances_sorted >= importances.mean()]  # trick to print just some most important features
-                                                  
-    top_10_features = importances_sorted_toplot[:20]
-    # Draw a horizontal barplot of importances_sorted
-    top_10_features.plot(kind='barh', color='lightgreen')
-    plt.title('Features Importances')
-    plt.show()
-    return importances_sorted_toplot
